separating handleClickHelper function from App
